requirements.txt
polars[rtcompat]>=0.20
pandas>=2.0
pyarrow>=15
requests>=2.31
python-dotenv>=1.0
scikit-learn>=1.4
xgboost>=2.0
shap>=0.45
scikit-optimize>=0.10
dash>=2.17
plotly>=5.20
